Feat: repo 어뎁터 해제 및 뷰모델 로직 이동
