+++ app/src/main/java/com/sopt/now/ui/search/SearchFragment.kt
@@ -1,22 +1,19 @@
 package com.sopt.now.ui.search
 
 import android.os.Bundle
-import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
-import com.google.gson.Gson
-import com.sopt.now.data.search.Repo
+import androidx.fragment.app.viewModels
 import com.sopt.now.databinding.FragmentSearchBinding
-import java.io.IOException
-import java.nio.charset.StandardCharsets
 
 class SearchFragment : Fragment() {
     private var _binding: FragmentSearchBinding? = null
     private val binding: FragmentSearchBinding
         get() = requireNotNull(_binding) { "FragmentSearchBinding is not initialized" }
 
+    private val searchViewModel: SearchViewModel by viewModels()
     private lateinit var repoAdapter: RepoAdapter
 
     override fun onCreateView(
@@ -31,7 +28,8 @@ class SearchFragment : Fragment() {
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
         setupAdapters()
-        loadJson()
+        setupObservers()
+        searchViewModel.loadJson(requireContext())
     }
 
     private fun setupAdapters() {
@@ -39,29 +37,15 @@ class SearchFragment : Fragment() {
         binding.rvRepo.adapter = repoAdapter
     }
 
-    private fun loadJson() {
-        val fileName = "fake_repo_list.json"
-        try {
-            context?.assets?.open(fileName)?.use { inputStream ->
-                val json = inputStream.readBytes().toString(StandardCharsets.UTF_8)
-                parseJson(json)?.let { repoAdapter.submitList(it) }
-            }
-        } catch (e: IOException) {
-            Log.e("SearchFragment", "Error loadJson", e)
-        }
-    }
-
-    private fun parseJson(jsonString: String): List<Repo>? {
-        return try {
-            Gson().fromJson(jsonString, Array<Repo>::class.java).toList()
-        } catch (e: Exception) {
-            Log.e("SearchFragment", "Error parseJson", e)
-            emptyList()
+    private fun setupObservers() {
+        searchViewModel.repoList.observe(viewLifecycleOwner) { repoList ->
+            repoAdapter.submitList(repoList)
         }
     }
 
     override fun onDestroyView() {
         super.onDestroyView()
         _binding = null
+        binding.rvRepo.adapter = null
     }
 }
